`tools/generator-go-sdk`: moving `generator` into the `internal` folder
