+++ tools/generator-go-sdk/internal/generator/data.go
