+++ tools/generator-go-sdk/internal/generator/helpers.go
