+++ tools/generator-go-sdk/internal/generator/helpers_test.go
