+++ tools/generator-go-sdk/internal/generator/object_definition.go
