+++ tools/generator-go-sdk/internal/generator/service.go
@@ -12,7 +12,6 @@ import (
 	"strings"
 
 	"github.com/hashicorp/pandora/tools/sdk/resourcemanager"
-
 	"github.com/hashicorp/pandora/tools/sdk/services"
 )
