+++ tools/generator-go-sdk/internal/generator/settings.go
