+++ tools/generator-go-sdk/internal/generator/stage_clients.go
