+++ tools/generator-go-sdk/internal/generator/stage_ids.go
