+++ tools/generator-go-sdk/internal/generator/stage_methods.go
