+++ tools/generator-go-sdk/internal/generator/stage_models.go
