+++ tools/generator-go-sdk/internal/generator/stage_predicates.go
