+++ tools/generator-go-sdk/internal/generator/stage_readme.go
