+++ tools/generator-go-sdk/internal/generator/stage_version.go
