+++ tools/generator-go-sdk/internal/generator/templater_clients_test.go
