+++ tools/generator-go-sdk/internal/generator/templater_constant.go
