+++ tools/generator-go-sdk/internal/generator/templater_constant_test.go
