+++ tools/generator-go-sdk/internal/generator/templater_constants_test.go
