+++ tools/generator-go-sdk/internal/generator/templater_id_parser_tests.go
