+++ tools/generator-go-sdk/internal/generator/templater_meta_client.go
