+++ tools/generator-go-sdk/internal/generator/templater_methods.go
