+++ tools/generator-go-sdk/internal/generator/templater_methods_discriminators_test.go
