+++ tools/generator-go-sdk/internal/generator/templater_models_discriminators_test.go
