+++ tools/generator-go-sdk/internal/generator/templater_models_test.go
