+++ tools/generator-go-sdk/internal/generator/templater_predicates.go
