+++ tools/generator-go-sdk/internal/generator/templater_readme.go
