+++ tools/generator-go-sdk/internal/generator/templater_readme_test.go
