+++ tools/generator-go-sdk/internal/generator/templater_version.go
